+++ sdk/test/configuration/yaml_trace_test.cc
@@ -1197,6 +1197,63 @@ tracer_provider:
   EXPECT_DOUBLE_EQ(visitor.ratio, 0.25);
 }
 
+TEST(YamlTrace, composable_probability_sampler_ratio_too_small)
+{
+  std::string yaml = R"(
+file_format: "1.0-trace"
+tracer_provider:
+  processors:
+    - simple:
+        exporter:
+          console:
+  sampler:
+    composite/development:
+      probability:
+        ratio: -0.5
+)";
+
+  auto config = DoParse(yaml);
+  ASSERT_EQ(config, nullptr);
+}
+
+TEST(YamlTrace, composable_probability_sampler_ratio_nan)
+{
+  std::string yaml = R"(
+file_format: "1.0-trace"
+tracer_provider:
+  processors:
+    - simple:
+        exporter:
+          console:
+  sampler:
+    composite/development:
+      probability:
+        ratio: nan
+)";
+
+  auto config = DoParse(yaml);
+  ASSERT_EQ(config, nullptr);
+}
+
+TEST(YamlTrace, composable_probability_sampler_ratio_too_large)
+{
+  std::string yaml = R"(
+file_format: "1.0-trace"
+tracer_provider:
+  processors:
+    - simple:
+        exporter:
+          console:
+  sampler:
+    composite/development:
+      probability:
+        ratio: 1.5
+)";
+
+  auto config = DoParse(yaml);
+  ASSERT_EQ(config, nullptr);
+}
+
 TEST(YamlTrace, composable_rule_based_sampler)
 {
   std::string yaml = R"(